wip 4 - first refactor
